+++ nodes/YandexCloudLockbox/resources/index.ts
@@ -1,0 +1,3 @@
+export * from './secret.operations';
+export * from './version.operations';
+export * from './payload.operations';